refactored function names, separate files for components and interfaces, passing props via interface
