+++ tabs/src/components/Tab.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { Scheduler } from "./BeaverHours/Scheduler";
+import { ScheduleForm } from "./BeaverHours/Scheduler";
 
 export default function Tab() {
   return (
     <div>
-      <Scheduler />
+      <ScheduleForm />
     </div>
   );
 }
